requirements.txt
weaviate-client>=3.26.7
pymongo>=4.5.0
streamlit>=1.29.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
